requirements.txt
torch
torchvision
pillow
transformers
